Initial import
Imports the `elliptic-curve` crate previously located at:

https://github.com/RustCrypto/signatures/tree/b46bbd95/elliptic-curve-crate

The `p256` code has been omitted both to give credit to @str4d for it
and to factor it into a separate crate.

No Travis CI configuration is included, with the goal of adding it in a
follow-up commit to ensure it passes before merging.