Add `AffineCoordinates` trait + DRY out point serialization
Adds a trait for obtaining the x and y-coordinates of an affine point.

So as to avoid exposing field elements as part of the public API, the
coordinates are first serialized as bytes.

Using this trait, it's possible to generically implement compressed and
uncompressed point serialization according to the SEC-1
Elliptic-Curve-Point-to-Octet-String encoding.

This commit also modifies the trait bounds of `FixedBaseScalarMul` to
bound on the `AffineCoordinates` trait, allowing the result to be an
affine point for the curve which can be passed as input to the newly
added `from_affine_point` methods on `CompressedCurvePoint` and
`UncompressedCurvePoint`.