Add `weierstrass::GenerateSecretKey` trait
Adds a `SecretKey` generation trait for Weierstrass curves supporting
curve-specific logic.

The trait is implemented for the `k256` and `p256` crates using a trial
generation method ("generate-and-pray") which attempts to decode a
randomly generated bytestring as a `Scalar` and loops until it succeeds.

This approach is suboptimal as it's non-constant-time, however this
should not pose a practical security concern as the timing sidechannel
shouldn't reveal any useful information (so long as RNG outputs do not
reveal information about prior or subsequent RNG outputs, which is a
property required for any secure CSPRNG).

It would be good to replace this with a truly constant-time approach
(e.g. using a modular reduction to ensure the scalar is in range,
rather than brute force)