Re-export types from `arithmetic` module at toplevel
This commit removes the `arithmetic` module from the public API of the
`k256` and `p256` crates, re-exporting the `AffinePoint`,
`ProjectivePoint`, and `Scalar` types from the crate's toplevel.

Additionally, it moves the `test_vectors` module to the toplevel,
gated under the `test-vectors` feature (or `cfg(test))`).

To make it clear these types are available conditionally based on the
`arithmetic` and `test-vectors` features, `doc_cfg` is used to tag
them in rustdoc as rendered by https://docs.rs