k256: add `Scalar::is_high`
Adds a method for determining if a scalar is greater than or equal to
the modulus divided by 2 (in constant time).

The ultimate purpose of this method is implementing non-malleable ECDSA
signatures as described in BIP 0062: Dealing with Malleability,
namely ensuring signatures are low-S normalized.

In the event the S-component of an ECDSA signature is "high", it should
be normalized by subtracting it from the modulus.