[WIP] Add `legacy` feature with ECDSA-related traits
ECDSA signing computes the `r` component of a signature by reducing the
affine x-coordinate of an ephemeral point mod n.

This adds a `ReduceAffineX` trait for performing this reduction, with
the goal of facilitating an ECDSA implementation which is generic over
(prime order) elliptic curves.

Additionally, it adds an `IsAffineYOdd` trait which is useful for
computing the "recovery ID" used to recover the public key which created
a signature (ala point compression).