k256: add ecdsa::recoverable::Signature::from_trial_recovery (#110)
The previous method was failing about 52% of the time, and was probably
failing to consider some combination of low-S or potentially r
overflowing the scalar field(?)

This commit a method for computing a `recoverable::Signature` from any
ECDSA/secp256k1 signature using trial recovery, and uses it in the
`RandomizedSigner<recoverable::Signature>` impl.

This method "brute forces" the recovery ID by trying both 0 or 1 until
the correct PublicKey is recovered, and returning an error if both fail.

It seems to have at least corrected the previous signature recovery
failures, although could probably be better optimized in the case of the
`RandomizedSigner` impl. This is noted in a TODO comment.