p521: remove vaporware crate
This crate is completely empty. Until we at least define a baseline set
of types for it (ala how the `p384` crate works presently) it doesn't
make sense to keep it around: it just adds additional pointless
maintenance/documentation work and misleads people into thinking it
might actually be useful.

This commit removes it.