k256: add `expose-field` feature; fix/CI benches (#161)
The benchmarks want direct access to `FieldElement` for benchmarking
purposes. This is a legitimate use case, but breaks encapsulation in
that we'd like to generally prevent exposure of `FieldElement` except
for "hazmat"-style use cases.

This is a well-recognized problem with Rust in general:
https://github.com/rust-lang/cargo/issues/2911

This commit adds a semi-hidden `expose-field` feature in order to fix
the benchmark. Perhaps it will be abused, but the alternative is not
having a benchmark, which seems bad too.

It also adds a CI build step to ensure the benchmarks compile to prevent
this sort of regression in the future.