k256/p256: bump `ff` and `group` deps to v0.8; MSRV 1.44+
Previously the `elliptic-curve` crate was depending on pre-releases of
`ff` and `group` sourced via git. It was upgraded to use the final
releases in RustCrypto/traits#292.

Because these crates depend on `bitvec` and its transitive dependencies,
an `arithmetic` feature was added to `elliptic-curve` in
RustCrypto/traits#293.

This commit upgrades the `k256` and `p256` crates to use the final
releases of `ff` and `group`. This ended up actually being a relatively
straightforward upgrade because we were already running off the latest
versions sourced from git.

The major notable change is `PrimeField::ReprBits` and the `to_le_bits`
and `char_le_bits` methods for obtaining `bitvec`-friendly types for
representing field elements.