k256: recoverable::Signature::recover_verify_key_from_digest_bytes (#205)
Adds a `recover_verify_key_from_digest_bytes` method which operates on a
byte array that's allegedy the digest of the message used to produce a
given signature.

Where `*Verifier` methods carry the risk of a signature forgery if they
operate on something that's not a message digest (since this violates
ROM under which ECDSA derives its security), there isn't a similar risk
for public key recovery: the best an attacker can do is force recovery
of a bogus key, and all recovered public keys are inherently untrusted
until they can be validated against a key fingerprint. So this method
doesn't have similar misuse concerns as signature verification.