k256: enable endomorphisms by default
The reason this wasn't enabled-by-default before was due to a lingering
US patent. However, as noted on #211, that patent just expired 🎉

The `bitcoin-core/secp256k1` library is also pursuing a full switch to
endomorphisms: https://github.com/bitcoin-core/secp256k1/pull/826

For now this commit just switches on the feature by default, but in the
future we can eventually rip out the non-endomorphism code and remove
the feature entirely.