k256+p256: fix DigestPrimitive feature gating
The previous feature gating prevented a `DigestPrimitive` impl from
being added in the event that the `sha256` feature was enabled, but the
`ecdsa` feature was not.

This is still a useful feature combination as it can still be used by
external signers even when just the `ecdsa-core` feature is enabled.