k256: use `crypto-bigint` for `scalar::MODULUS`
This commit begins the process of starting to retrofit representing the
scalar modulus as a `crypto_bigint::UInt`.

It isn't a perfect fit in the current form as the `force-32-bit` cargo
feature complicates things. It would probably make sense to get rid of
it in the next breaking release, especially as it isn't additive.
Instead it should be easy to do 32-bit testing with `cross`.

All that said, it's now possible to calculate constants like
`FRAC_MODULUS_2` at compile time using `const fn` which is pretty neat.