k256: remove `force-32-bit` feature
This feature was originally added for local testing purposes, but
greatly complicates interactions with additional crates like
`crypto-bigint` which are gated entirely on the target's word size.

The testing workflow this provided can still be accomplished using a
tool like `cross` by testing against a 32-bit target:

```
$ cross test --target powerpc-unknown-linux-gnu --release --all-features
```

Bonus points for testing against a big endian target to ensure there
aren't any endian-related bugs as well.

Breaking change in that it removes a feature.