Bump `elliptic-curve` to v0.11.0-pre (#402)
Sourced from git.

This includes transitive updates to ff and group v0.11. A major notable
change of these updates is switching several APIs to use
`subtle::Choice` and `subtle::CtOption` instead of `bool`/`Option`.

Fortunately, the crates in this repo are already written with
constant-time implementations internally, so this was an easy change.