p256: use `U256` as the internal representation of `Scalar`
This is a minimally invasive change to use `crypto_bigint::U256` as the
internal representation of `p256::Scalar`.

There are several places that various functions are still using a
`U64x4` type which could be made to work over 8 x 32-bit limbs instead
on 32-bit platforms, which is what `U256` uses internally.

This commit doesn't attempt to implement those changes, but rather just
switch out the type used for the internal representation. This should
hopefully avoid any regressions in this particular change.