k256: factor `mul_shift_var(time)` into `wide` module
This, combined with a small fixup to `conditional_add_bit`, ensures that
all of the `Scalar`-related word size-specific gating is relegated to
the `wide32.rs` vs `wide64.rs` modules.

Ideally these could be replaced with a common `U512`-based
implementation which further reduces this gating, but that's left for a
followup step.