k256+p256: handle identity point in `GroupEncoding`
`GroupEncoding` uses a fixed-width `Repr` whereas SEC1 is variable-width
with respect to compressed points versus the identity.

This change allows 33-bytes of zeroes to be used as the identity.
Technically that's not a valid SEC1 encoding: the SEC1 encoding for the
identity is 1-byte: 0x00. However, this is the best we can do with a
fixed-width encoding.

See also: #443 and #444