Update `ecdsa` crate and leverage generic ECDSA implementations
This updates the crates in this repo to leverage the following generic
implementations recently added to the `ecdsa` crate:

- Sign: RustCrypto/signatures#396
- Verify: RustCrypto/signatures#397

The `p256` crate is able to directly utilize the generic implementation
from the `ecdsa` crate.

The `k256` crate retains its previous custom implementation, which
provides a secp256k1-specific take on public key recovery, but can also
directly leverage optimizations for computing linear combinations, which
there are not presently traits in `elliptic-curve` or `group` to
express.