p256: make `AffinePoint::generator` closer to const-friendly
This is a stepping stone towards solving #508.

Replaces the implementation of `AffinePoint::generator` with one that
uses Rust literal syntax to construct the field elements which represent
the base point's affine coordinates.

Coordinates are specified in canonical form for easy auditability, then
converted to Montgomery form. With `const fn`, this conversion could
happen completely at compile time.

This can't yet be made fully `const` though, as there isn't a `const fn`
constructor for `subtle::Choice`, which is used to represent the point
at infinity.