k256: fallible `AffinePoint` to `VerifyingKey` conversion
The inner type of `VerifyingKey` is a `PublicKey`, which is itself a
wrapper for an inner curve-specific `AffinePoint` which ensures that the
inner point is *NOT* the additive identity / point at infinity.

Therefore, any conversion from an `AffinePoint` is potentially fallible,
so the existing `From` conversions need to be changed to `TryFrom`.

This commit also adds some additional type conversions between
`AffinePoint`/`ProjectivePoint` and `PublicKey`/`VerifyingKey`, and also
includes them in `p256` for consistency.