k256: use SHA-256 for computing RFC6979 for `recoverable::Signature`
For whatever reason when the Ethereum developers decided to use
Keccak256 as the digest function for ECDSA itself, they still used
SHA-256 as the digest function for RFC6979, possibly because they didn't
realize they needed to change the digest function in two places.

Previously the implementation in `k256` would use the same digest
function for both (typically Keccak256) but we've had various complaints
that doing so doesn't match the test vectors produced by various other
libraries (though the signatures will still verify).

This commit uses HMAC-DRBG instantiated with SHA-256 for computing
RFC6979 for `ecdsa::recoverable::Signature`, regardless of what digest
function is used to hash the message. In theory this should match the
Ethereum test vectors.