p256: use `impl_field_element!` macro
Uses the same macro as `p384` to define the `FieldElement` type.

Isolates core arithmetic functionality in a `p256_64.rs` module, with
the goal of eventually producing a 32-bit implementation as well.

Also adds a `p256_32.rs` module which provides a skeleton of a 32-bit
implementation, although most of the functionality is currently
implemented in terms of 64-bit integers, with the goal of facilitating
an incremental conversion to a 32-bit backend.