p224: initial crate
Initial crate boilerplate with a `NistP224` type, `Curve` impl, and
type aliases for the generic types which can work with that.

Like P-521, P-224 has a somewhat odd size that doesn't map well to
64-bit limbs.

To make 64-bit limbs work, it uses `U256` as the `Curve::UInt` type,
however like P-521 (which uses `U576`) this means `FieldBytes` is the
wrong size.

This will require upstream changes in the `elliptic-curve` crate to
correct, and also for now complicates any further work on the
implementation.