k256: add complete end-to-end Ethereum signature recovery test
Adds a test which signs a message using Keccak256 as the digest
function, testing that the resulting signature matches the one produced
by web3.js, as used in the ethers-rs test vectors.

After the signature has been produced, it recovers the `VerifyingKey`
from the message digest and signature, testing that it's the correct
`VerifyingKey`, and also testing that it successfully verifies the
signature.