p224: `arithmetic` feature (#815)
Now that #813 has landed it seems prudent to add a real `arithmetic`
feature.

This adds a feature similar to the `p256` and `p384` crates which
exposes the following types which provide a curve arithmetic
implementation:

- `AffinePoint`
- `ProjectivePoint`
- `Scalar`

The `wip-arithmetic-do-not-use` feature is now removed as well. While
technically SemVer breaking, it wasn't supposed to be used in the first
place!