bp384: `CurveArithmetic` + `PrimeCurveParams`
Adds impls of the `CurveArithmetic` and `PrimeCurveArithmetic` traits,
the latter of which defines the coefficients of the curve equation as
well as the coordinates of the generator point.

Impls have been added for both `Brainpool384r1` and `Brainpool384t1`,
which use different curve equation coefficients and generators, but
share the base and scalar field implementations.