p521: fix projective arithmetic tests
The tests added in #950 weren't actually running due to feature gating
(the feature is named `wip-arithmetic-do-not-use`, but the conditional
gating was using `arithmetic` instead).

It turns out a small modification to `primeorder` was needed, since the
`impl_projective_arithmetic_tests` macro was relying on `Into`
conversions from core arrays to `GenericArray` which top out at
64-element arrays, and `p521`'s field elements are 66-bytes.

With the modification in place, the tests pass!