p521: remove `pub` from `arithmetic` module (#961)
The relevant types from the `arithmetic` module are available via
re-exports, so direct access to the module is unnecessary. It was
accidentally exposed in the v0.13.1 release (#958).

Exposing the module directly is inconsistent with p192/p224/p256/p384,
so this removes it. This is technically a breaking change, but it was
added in v0.13.1 which was released a few minutes ago, and will be
yanked after this is released.