p521: merge `u576_to_le_bytes` into `FieldBytes::from_uint_unchecked`
Moves `u576_to_le_bytes` from the `util` module to inside of
`from_uint_unchecked`.

The old function in the `util` module was not general-purpose, but
specialized to the case of decoding P-521 base field elements in the
unsaturated format used by fiat-crypto's Solinas prime backend.

Additionally `FieldElement::from_uint_unchecked` is the only caller, and
it really is an implementation detail of that function.