primefield: crate boilerplate
Right now the `primeorder` crate, which contains elliptic curve addition
formulas, is also curiously overloaded with some macros for generating
field implementations, even though those have nothing in particular to
do with the curve formulas.

This is largely because `primeorder` was an existing dependency of the
other crates, so it was a convenient place to put it.

This begins the process of extracting it out into a separate crate. It
should also make it easier to perform the `fiat-crypto` upgrade, since
we can put the new implementation in `primefield` and leave the existing
one in `primeorder` until all crates have been migrated.