Restore "p521: fiat-constify update (#1003)"
This reverts commit 26be150c5c79b93f65f9b522e1626ed9426b4305.

Now that we've completed the `elliptic-curve` v0.14.0-pre upgrade, we
can restore the fiat-crypto upgrade.

This commit takes a slightly different approach and puts the macro
implementation for the newest version of `fiat-crypto` into the
new `primefield` crate which was added in #1013.

The goal will be to update all of the crates and then remove the old
macros from `primeorder` entirely.