Add `EcdsaCurve` trait impls
This is a marker trait added in RustCrypto/signatures#787 which marks an
elliptic curve as being intended for use with ECDSA.

It's the intended successor of the `SignPrimitive` and `VerifyPrimitive`
traits, carrying with it the one bit of information needed to eliminate
the need for custom impls: a flag for low-S normalization.

`bp256`, `bp384`, `k256`, `p192`, `p224`, `p256`, `p384`, and `p521`
have all received trait impls.

Of those, only `k256` has the `NORMALIZE_S` flag set to `true`.