p384: remove default `ecdh` feature (#1112)
This default feature is inconsistent with all of the other curve
implementations in this repo, which don't enable the `ecdh` feature by
default.

We could go either way, enabling it in all the other curves, or just
disabling it here. This PR opts to do the latter, as it's simpler.

Closes #1047