k256: use `std::sync::LazyLock` instead of `once_cell`
When building with the `precomputed-tables` and `std` features, uses the
built-in implementation of `LazyLock` in lieu of `once_cell::sync::Lazy`
thus avoiding the extra crate dependency.

The `critical-section` feature is retained and uses
`once_cell::sync::Lazy`.

When `precomputed-tables`, `critical-section`, and `std` are all
enabled, uses `once_cell/critical-section` and enables `once_cell/std`,
which should use a `std::sync::Mutex` for synchronization.