primefield+p384: upgrade to `fiat-crypto` v0.1.4 (#1157)
Updates the field element type generating macros in the `primefield`
crate to be compatible with the latest `fiat-crypto` code as
postprocessed by `fiat-constify`, and regenerates the `p384` field impls
using `fiat-crypto` v0.1.4.

This update leverages `const_mut_refs` to be much closer to the upstream
`fiat-crypto` output, while still providing `const fn` support.

This change also starts extracting a `fiat` module within `primeorder`
for macros that are specific to `fiat-crypto` output.

With this upgrade in place, upgrading the rest of the curve crates
should be comparatively straightforward.