primefield: fix modulus bit size calculation (#1176)
Calculates the bit size of the modulus based on the byte length of
`FieldBytes` (as opposed to the size of a `Uint`, which may be larger
than the bit length of the modulus depending on e.g. platform word size)

This makes it possible to use the full `fiat_field_arithmetic!` macro
which implements inversions for `p224`, where the bit size was
previously being miscomputed on 64-bit platforms.

Since it's no longer needed, this merges the
`fiat_field_arithmetic_core!` macro back into `fiat_field_arithmetic!`.