p256: begin using `primefield`
This represents an initial attempt to convert some of the functionality
of `FieldElement` to using macros from the `primefield` crate.

Changes have been made carefully and minimally along with benchmarking
to ensure this does not impact performance (unlike the last attempt to
do this, see #885)

Trying to use the full `field_element_type!` did appear to result in a
benchmark regression, so this instead extracts a
`field_element_type_core!` macro (again) and begins to use it in `p256`,
ensuring that there are no benchmark regressions.