p256: use `primefield` macros (#1180)
Write most of the `FieldElement` implementation in `p256` using the same
macros we're applying to other crates.

This somewhat mysteriously seems to cause the following performance
regression on a macOS/M2 target:

    point operations/point-scalar mul
                        time:   [155.25 µs 155.60 µs 156.13 µs]
                        change: [+2.3080% +2.7473% +3.4111%] (p = 0.00 < 0.05)
                        Performance has regressed.

Drilling down on the issue, it seems related to the `#[inline]`
directives on the core arithmetic trait impls:

https://github.com/RustCrypto/elliptic-curves/blob/f259ec4/primefield/src/lib.rs#L511-L529

It would be good to confirm this is a real issue, possibly through
detailed codegen examination, though the total size of involved code is
relatively large.

However, the change is small enough to be within a range that is
difficult to benchmark, and it's not yet clear how other changes
(e.g. switching to a `crypto-bigint` backend) will impact performance.

So this commit accepts this performance regression for now with a note
to follow up on potentially switching inlining off on the `core::ops`
impls to see if that addresses the issue if other code changes don't
make this irrelevant.