Expose `AffineCoordinates::y`
Companion PR to RustCrypto/traits#1891 which adds access to affine
y-coordinates.

Notably we've just added an Edwards curve with `ed448-goldilocks`, and
these curves use compressed Edwards y-coordinates, so y-coordinate
access is needed to even implement point compression generically for
these curves.

Note that due to RustCrypto/elliptic-curves#1229 it's currently not
possible to impl the extened `AffineCoordinates API for the Decaf
implementation in the `ed448-goldilocks` crate. This PR comments out the
relevant code for now with a TODO to fix it.