ed448-goldilocks: source `zeroize` from `elliptic-curve`
The `zeroize` crate is a hard dependency of `elliptic-curve` (which uses
its traits in bounds), which is in turn a hard dependency of
`ed448-goldilocks`.

As such, a `zeroize` feature does not accomplish much, since `zeroize`
is still a transitive dependency.

This commit removes the `zeroize` feature and explicit `zeroize`
dependency and sources it instead as `elliptic_curve::zeroize`.