primefield: remove `shr` support
It was noted as buggy in #1319 because `shr` was being computed on
values within the Montgomery domain directly. They either need to be
converted to canonical form first, or a multiplication needs to be used
in place of a bit shift.

Turns out it was completely unused, aside from fulfilling a bound in the
`elliptic-curve` crate, which has since been removed.

This removes `shr` support from all curves whose `Scalar` used an
internal Montgomery representation. It has been retained on `k256` and
`p256` which use canonical form `Scalar` types.

Closes #1319