bignp256: rename from `bign256` (#1430)
According to https://apmi.bsu.by/assets/files/std/bign-spec294.pdf
the correct name for the curve is `bign-curve256v1`.

The name `bignp256` follows the somewhat loose convention of crates in
this repo which use abbreviations for the names.

`*p256` denotes it's a 256-bit prime order curve, following the NIST
"P-256" notation. `bp256` uses this, for example, though that's an
abbreviation for a curve (or rather, two curves) named `brainpoolP256*`.

Rename discussed in #796.