primefield: make exponent of `pow_vartime` a `Uint`
Previously it was `&[u64]`, which dates back to the original code which
was adapted from `ff_derive`.

This changes the argument to be a `Uint` (which can be a different
number of limbs from `Self`), and also changes `MontyFieldParams::T` to
be a `Uint` as well (where `T` can be used to compute many other
constants, often by raising them to the power of `T`).

This simplifies dealing with 32-bit vs 64-bit platforms, and also should
make it easier to provide a generic implementation of `sqrt` where we
may want to calculate constants based on `T`.