primefield: generic implementation of `sqrt` (#1446)
Adapts the implementation of `sqrt` from the `ff_derive` crate:

https://github.com/zkcrypto/ff/blob/f77b814/ff_derive/src/lib.rs#L509

...which is in turn adapted from this paper:

https://eprint.iacr.org/2012/685.pdf

For the most part the sqrt impls replaced are special case versions of
the code which would've otherwise been derived by `ff_derive`.

Some TODOs are left to make more computations `const`, and notably
this removes usages of Atkins algorithm, which should also be added
back in a followup.