p521: add 32-bit fiat scalar backend; use `primefield`
Synthesizes a 32-bit backend using fiat-crypto v0.1.5. This was
previously a bit tricky because it causes a stack overflow with default
ulimits. Whether this is actually a good idea is a bit questionable as
it ends up being quite large at 22.5 KLOCs.

The particularly nice thing it does afford, however, is being able to
use `primefield` to write the field implementation. This makes it
consistent with the other implementations and addresses some TODOs, like
a sqrt implementation.