p521: use `primefield` for scalar field
We attempted to synthesize a fiat-crypto 32-bit scalar backend in #1465.
It was 22.5 KLOCs, took 45s to run rustfmt over, and rustc spent longer
than I was willing to wait trying to compile it without `--release`
(several minutes).

In lieu of that, this stubs out a 32-bit backend so we can use
`primefield`, but implements it by wrapping the 64-bit backend and
applying the same sorts of conversions we were already using, just doing
it in stubbed out `fiat-crypto` shaped code where we could conceivably
drop-in a 32-bit backend if rustc didn't choke on it.